Scopus search

Scopus search

scopus_search(
  query = NULL,
  count = 25,
  start = 0,
  type = "search",
  search_type = "scopus",
  facets = NULL,
  view = NULL,
  date = NULL,
  sort = NULL,
  content = NULL,
  subj = NULL,
  key = NULL,
  ...
)

scopus_search_loop(
  query = NULL,
  count = 25,
  start = 0,
  type = "search",
  search_type = "scopus",
  facets = NULL,
  view = NULL,
  date = NULL,
  sort = NULL,
  content = NULL,
  subj = NULL,
  key = NULL,
  ...
)

Arguments

query
(character) query terms, as a single character vector

count
(integer/numeric) results to return: default: 25

start
(integer/numeric) offset value, default: 0

type
(character) type of search, default: search

search_type
(character) search type, default: scopus

facets
(list) facets, see https://dev.elsevier.com/tecdoc_api_facets.html for how to construct facet queries

view
the fields to return, see https://dev.elsevier.com/guides/ScopusSearchViews.htm

date
Represents the date range associated with the search, with the lowest granularity being year. e.g. 2002-2007

sort
Represents the sort field name and order. A plus in front of the sort field name indicates ascending order, a minus indicates descending order. If sort order is not specified (i.e. no + or -) then the order defaults to ascending (ASC). Up to three fields can be specified, each delimited by a comma. The precedence is determined by their order (i.e. first is primary, second is secondary, and third is tertiary). . e.g., "overDate,-title"

content
filter specific categories of content that should be searched/returned. one of: core, dummy, all (default)

subj
the subject area code associated with the content category desired. Note that these subject code mapping vary based upon the environment in which the request is executed. See Details for choices.

key
(character) api key. get a key at https://dev.elsevier.com/index.html

...
curl options passed on to crul::HttpClient

Details

Rate limits for search are 20,000 per every 7 days. You likely won't make that many requests in 7 days, but if you do e.g., make 20K in 5 days, then you have to wait 2 days for the clock to reset, than you'll be able to make 20K again.

See https://dev.elsevier.com/api_key_settings.html for rate limit information.

See https://dev.elsevier.com/sc_search_tips.html for help/tips on searching

subj choices include

AGRI: Agricultural and Biological Sciences

ARTS: Arts and Humanities

BIOC: Biochemistry, Genetics and Molecular Biology

BUSI: Business, Management and Accounting

CENG: Chemical Engineering

CHEM: Chemistry

COMP: Computer Science

DECI: Decision Sciences

DENT: Dentistry

EART: Earth and Planetary Sciences

ECON: Economics, Econometrics and Finance

ENER: Energy

ENGI: Engineering

ENVI: Environmental Science

HEAL: Health Professions

IMMU: Immunology and Microbiology

MATE: Materials Science

MATH: Mathematics

MEDI: Medicine

NEUR: Neuroscience

NURS: Nursing

PHAR: Pharmacology, Toxicology and Pharmaceutics

PHYS: Physics and Astronomy

PSYC: Psychology

SOCI: Social Sciences

VETE: Veterinary

MULT: Multidisciplinary

Examples

if (FALSE) {
res <- scopus_search(query = "ecology")
res

#scopus_search(query = x, type = "abstract")

# looping through
res <- scopus_search_loop(query = "ecology community elk cow")

# using facets
## scopus_search
# res <- scopus_search(query = "ecology", facets = "subjarea(count=5)")
# res
# res$`search-results`$link
# res$`search-results`$entry
# res$`search-results`$facet

## more examples
# x <- scopus_search(query = "ecology", facets = "language(count=4)",
#   count = 1)
# x$`search-results`$facet
# x <- scopus_search(query = "ecology",
#   facets = "pubyear(count=3);doctype();language(count=4)")
# x$`search-results`$facet

## scopus_search_loop
# res <- scopus_search_loop(query = "ecology", facets = "subjarea(count=5)",
#   count = 200)
# res$found
# head(res$results)
# NROW(res$results)
# res$facets

# sort
x <- scopus_search(query = "ecology", sort = "-title")
}